[test] move test to a file requiring the full stdlib
- SIMD types are not available in the minimal stdlib
